[DEV-35] Move Filters to a seperate file and add apply button
