+++ packages/chart/src/index.html
@@ -28,7 +28,8 @@
       <option value="Hispanic or Latino">Test 2</option>
     </select> -->
 
-  <div class="react-container" data-config="/examples/private/yaxis-test.json"></div>
+  <div class="react-container" data-config="/examples/private/filters.json"></div>
+  <!-- <div class="react-container" data-config="/examples/private/yaxis-test.json"></div> -->
   <!-- <div class="react-container" data-config="/examples/dynamic-legends.json"></div> -->
   <!-- <div class="react-container" data-config="/examples/cutoff-example-config.json"></div> -->
   <!-- <div class="react-container" data-config="/examples/covid-confidence-example-config.json"></div> -->
